remove use of  tt tag in section headings
The tt tag was applied to the &nbsp that connects the formatted section number with the section heading text. The purpose of this is unclear to me. Possibly the intent was to ensure more even alignment of heading text between successive sections. This edit removes the use of the <tt> tag, as it has been deprecated in html5. If a monospace font &nbsp is still needed, then it should be enclosed in a span with a class of "stt".